+++ include/MoveBG/MapObjTown.hpp
@@ -147,7 +147,7 @@ public:
 	THideObjInfo(const char* name = "オブジェ出現情報");
 
 	virtual void load(JSUMemoryInputStream&);
-	virtual void action(long);
+	virtual void action(s32);
 
 public:
 	/* 0x44 */ u32 unk44;
